Fixed issue when the flang in not installed
If the Frontend Language extension is configured with languages not available in the backend, the Lang::get() call would always return values that are configured in the backend. This is clearly not what we want, as we want to support any language in the frontend while not necessarly have to install the language in Symphony.

In the backend, it's always Lang::get() value that is used.

Re #11